requirements.txt
appdirs==1.4.4
application-utility==1.3.2
asgiref==3.3.1
asn1crypto==1.4.0
astroid==2.4.2
Beaker==1.11.0
btrfsutil==5.9
CacheControl==0.12.6
ceph==1.0.0
ceph-volume==1.0.0
cephfs==2.0.0
cephfs-shell==0.0.1
cffi==1.14.4
chardet==3.0.4
colorama==0.4.4
contextlib2==0.6.0.post1
cryptography==3.3.1
cupshelpers==1.0
decorator==4.4.2
distlib==0.3.1
distro==1.5.0
Django==3.1.5
django-embed-video==1.4.0
django-recaptcha==2.0.6
django-summernote==0.8.11.6
django-videofield==0.1.1
docopt==0.6.2
evdev==1.4.0
html5lib==1.1
idna==2.10
isort==5.7.0
keyutils==0.6
lazy-object-proxy==1.4.3
LibAppArmor==3.0.1
louis==3.16.0
lutris==0.5.8.1
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.2
netsnmp-python==1.0a1
npyscreen==4.10.5
numpy==1.19.5
ordered-set==4.0.2
packaging==20.8
pacman-mirrors==4.19.1
pep517==0.9.1
Pillow==8.1.0
ply==3.11
progress==1.5
pycairo==1.20.0
pycparser==2.20
pycups==2.0.1
pycurl==7.43.0.6
PyGObject==3.38.0
pylint==2.6.0
pylint-django==2.4.2
pylint-plugin-utils==0.6
pyopencl==2020.2.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyQt5==5.15.2
PyQt5-sip==12.8.1
pysmbc==1.0.23
python-libtorrent==1.2.10
pytools==2021.1
pytz==2020.5
PyYAML==5.3.1
rados==2.0.0
rbd==2.0.0
reportlab==3.5.59
requests==2.25.1
resolvelib==0.5.4
retrying==1.3.3
rgw==2.0.0
six==1.15.0
sqlparse==0.4.1
team==1.0
toml==0.10.2
udiskie==2.2.0
urllib3==1.26.1
webencodings==0.5.1
wrapt==1.12.1
